[#46] refactor: Exhibition aggregate -> Category aggreate 분리
