+++ backend/src/main/java/com/terria/sidedish/api/CategoryController.java
@@ -1,0 +1,5 @@
+package com.terria.sidedish.api;
+
+public class CategoryController {
+    
+}
