+++ backend/src/main/java/com/terria/sidedish/api/ExhibitionController.java
@@ -22,8 +22,8 @@ public class ExhibitionController {
     private final ExhibitionService exhibitionService;
 
     @ApiOperation(
-            value = "특정 기획전에 속한 데이터(카테고리, 반찬 카드) 조회",
-            notes = "특정 기획전에 속한 데이터(카테고리, 반찬 카드)를 조회한다.",
+            value = "특정 기획전에 속한 카테고리 조회",
+            notes = "특정 기획전에 속한 카테고리를 조회한다.",
             produces = "application/json",
             response = ExhibitionResponse.class
     )
