+++ backend/src/main/java/com/terria/sidedish/domain/entity/aggregate/Category.java
@@ -1,11 +1,7 @@
-package com.terria.sidedish.domain.entity.reference;
+package com.terria.sidedish.domain.entity.aggregate;
 
-import com.terria.sidedish.domain.entity.link.CategoryHasSideDish;
 import lombok.*;
 import org.springframework.data.annotation.Id;
-import org.springframework.data.relational.core.mapping.MappedCollection;
-
-import java.util.Set;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -16,16 +12,12 @@ public class Category {
     private String title;
     private long exhibitionId;
 
-    @MappedCollection(idColumn = "category_id")
-    private Set<CategoryHasSideDish> categoryHasSideDishes;
-
     @Override
     public String toString() {
         return "Category{" +
                 "id=" + id +
                 ", title='" + title + '\'' +
                 ", exhibitionId=" + exhibitionId +
-                ", categoryHasSideDishes=" + categoryHasSideDishes +
                 '}';
     }
 }
