+++ backend/src/main/java/com/terria/sidedish/domain/entity/aggregate/SideDish.java
@@ -1,6 +1,7 @@
-package com.terria.sidedish.domain.entity.reference;
+package com.terria.sidedish.domain.entity.aggregate;
 
 import com.terria.sidedish.domain.entity.link.SideDishHasDiscountEvent;
+import com.terria.sidedish.domain.entity.reference.SideDishImage;
 import lombok.*;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.MappedCollection;
