+++ backend/src/main/java/com/terria/sidedish/dto/response/CategoryResponse.java
@@ -1,13 +1,11 @@
 package com.terria.sidedish.dto.response;
 
-import com.terria.sidedish.domain.entity.reference.Category;
+import com.terria.sidedish.domain.entity.aggregate.Category;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
-
-import java.util.List;
 
 @ApiModel
 @Getter
@@ -21,10 +19,7 @@ public class CategoryResponse {
     @ApiModelProperty(required = true, value = "카테고리 타이틀")
     private String categoryTitle;
 
-    @ApiModelProperty("반찬 카드 목록")
-    private List<SideDishCardResponse> sideDishCardResponses;
-
-    public static CategoryResponse from(Category category, List<SideDishCardResponse> sideDishCardResponses) {
-        return new CategoryResponse(category.getId(), category.getTitle(), sideDishCardResponses);
+    public static CategoryResponse from(Category category) {
+        return new CategoryResponse(category.getId(), category.getTitle());
     }
 }
