+++ backend/src/main/java/com/terria/sidedish/dto/response/ExhibitionResponse.java
@@ -1,5 +1,6 @@
 package com.terria.sidedish.dto.response;
 
+import com.terria.sidedish.domain.entity.aggregate.Exhibition;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.AllArgsConstructor;
@@ -23,7 +24,7 @@ public class ExhibitionResponse {
     @ApiModelProperty("카테고리 목록")
     private List<CategoryResponse> categoryResponses;
 
-    public static ExhibitionResponse from(long exhibitionId, String exhibitionTitle, List<CategoryResponse> categoryResponses) {
-        return new ExhibitionResponse(exhibitionId, exhibitionTitle, categoryResponses);
+    public static ExhibitionResponse from(Exhibition exhibition, List<CategoryResponse> categoryResponses) {
+        return new ExhibitionResponse(exhibition.getId(), exhibition.getTitle(), categoryResponses);
     }
 }
