+++ backend/src/main/java/com/terria/sidedish/repository/SideDishRepository.java
@@ -1,6 +1,6 @@
 package com.terria.sidedish.repository;
 
-import com.terria.sidedish.domain.entity.reference.SideDish;
+import com.terria.sidedish.domain.entity.aggregate.SideDish;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
